pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scTimeBench"  # This is the name you use for 'pip install'
version = "0.1"
dependencies = [
    "PyYAML",
    "scanpy",
    "scikit_learn",
    "fastparquet",
    "pyarrow",
]

[tool.setuptools.packages.find]
where = ["src"]  # Tells Python: "The actual code lives inside the src folder"

[project.optional-dependencies]
test = [
    "pytest",
]
dev = [
    "pre-commit",
]
benchmark = [
    "networkx",
    "numpy",
    "geomloss",
    "pykeops",
    "scanpy[leiden]",
    "graphviz",
    "celltypist"
]
pypsupertime = [ "pypsupertime" ]
sceptic = [ "sceptic" ]

[project.scripts]
scTimeBench = "scTimeBench.main:main"

# --- UV SPECIFIC CONFIGURATION ---

[tool.uv]
# This tells uv that psupertime is part of your development 'workspace'
workspace = { members = ["extern/psupertime", "extern/sceptic"] }

[tool.uv.sources]
# This ensures uv looks in your submodule folder and treats it as an editable install.
pypsupertime = { editable = true, workspace = true }
sceptic = { editable = true, workspace = true }
